1.6.0 remove any ipfs gateway interaction functions in favour of https://github.com/emendir/IPFS-Toolkit-Python
