+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "robonomics-interface"
-version = "1.5.0"
+version = "1.6.0"
 description = "Robonomics wrapper over https://github.com/polkascan/py-substrate-interface created to facilitate programming with Robonomics"
 authors = ["Pavel Tarasov <p040399@outlook.com>"]
 license = "Apache-2.0"
