+++ robonomicsinterface/__init__.py
@@ -18,4 +18,3 @@ from .classes import (
     SubEvent,
     Subscriber,
 )
-from .ipfs_utils import web_3_auth, ipfs_upload_content, ipfs_get_content
