+++ robonomicsinterface/exceptions.py
@@ -23,11 +23,3 @@ class InvalidExtrinsicHash(Exception):
     """
 
     pass
-
-
-class FailedToUploadFile(Exception):
-    """
-    Failed to upload a file to Crust Network.
-    """
-
-    pass
